Frame Trapped should now work with stage quality set to low. Fixed small bug in the MemoryEditor. Added some more memory addresses.
The MemoryEditor had a small bug in it, we should now be able to read a byte array from memory.
Added some more memory addresses, these are:
GetCharacter(int player) - Returns the character name as a string use GetCharacter(1) for player 1 and GetCharacter(2) for player 2.
GetAnimationFrame(int player) - Gets the currently playing character animation frame.
GetPlayerScript(int player) - Gets the currently playing script.
GetComboCounter - Gets the combocounter as an int. Both players share the same combocounter! As long as it's obove 0 you are in a combo.